Changes to Environments 1 & 4 along with the addition of a new skybox for environment 1
Added a full glft for environment 4, but cannot figure out how to fix the navmesh path as making one on Blender 2.79 crashes it.

The whiteboards in environment 4 needs to be configured to set the canvas as the material "WhiteboardCanvas", that should work across both environments 1 & 4, also might be worth refactoring environment 2 and 3 since they are no longer being used. As well as this, I noticed some issues in the
Environment 4 GLTF which seems to load slower and experience lag when looking at the bean-bags in the red dome (the big one), not sure if its just the geometry or what the issue might be.

Also added rotation to the player as they were facing the wrong way in environment 4, this does not affect what they see, but if you spawn in environment 4, you will be most likely looking at a black wall otherwise.

Need to include some mechanism to teleport player and reorient them in the right way when switching environments, otherwise they just go off of the nav mesh